Query: move LabelVisitor to query package
LabelVisitor (and friends) are only used by BlazeQueryEnvironment, so
move them to live next to it, and reduce visibility to the package.

Also make a tiny refactoring in LabelVisitor to make the
VisitationAttributes inner class static.

This is in preparation for a rewrite of BlazeQueryEnvironment that
also requires changes to LabelVisitor.

PiperOrigin-RevId: 250129148