Java coverage: fix coverage when deploy jars end up on the classpath.
Handle the cases when a deploy jar is executed from a java program or when one or more deploy jars end up on the runtime classpath.

Also remove traces of the old variables referencing the 'new' implementation. The 'old' implementation is no longer in Bazel.

Make JacocoCoverageRunner compatible with JDK 9 (which made the system classloader no longer an instance of URLClassLoader).

Fixes #8472.

FYI: These cases are already handled in the internal JacocoCoverageRunner and this change syncs the two versions.

Closes #8357.

PiperOrigin-RevId: 250240091