C++: Flips --incompatible_do_not_split_linking_cmdline
GH issue: #7687

The flag is now true by default in Bazel.

Tested:
https://buildkite.com/bazel/bazel-at-head-plus-downstream/builds/996

Envoy is showing up as broken. Udp breakage is due to a different reason.

The Remote Execution is also fine to break with this change according to buchgr@. That project is using an old version of the default crosstool which still patches features from Bazel.

For Envoy I filed https://github.com/envoyproxy/envoy/issues/6968, suggested a fix and they will fix it soonish. If they don't, please disable Envoy downstream or I might miss the 0.27 cut.

RELNOTES:
PiperOrigin-RevId: 250252235