Fixed Python 3 incompatibility in make_rpm tool
This addresses #7914 , i.e. allowing the `pkg_rpm` rule to be used with either a PY2 or PY3 host configuration, by fixing a Python 3 incompatibility in `make_rpm.py`.

I added a test case that ensures the fixed codepath is covered. It effectively fakes out the `rpmbuild` executable with a shell script, so I'm not sure if there will be concerns with cross-platform compatibility. If so, I could create a Python script for this, or I'm open to other suggestions.

Closes #8429.

PiperOrigin-RevId: 250477353