Windows, launcher: support python/bash toolchain
Previously, we assume the python/bash binary path passed to the launcher is an absolute path specified by `--python_path` or `--shell_executable`, but this is not true when using python/shell toolchain. When using the python toolchain, the given python binary path will be a runfile path, we can use `Rlocation` to find the absolute path.

Fixes https://github.com/bazelbuild/bazel/issues/7947

Closes #8440.

PiperOrigin-RevId: 250569171