git.bzl: correctly quote shallow_since
...while still accepting that the parameter might be missing; in the
latter case, an empty argument was added, causing that invocation of
git to fail. Note, however, that the problem is of purely aesthetic
nature, as
- the empty argument would only be added if no information about
  shallow cloning can is available
- upon error, a retry without the shallow information would be made
  anyway to support older versions of git.
Nevertheless, let's not call processes in a way we know is bound to
fail; do this by moving the necessary quoting to the definition of
the "shallow" variable, rather than at its interpolation into the
shell command.

Fixes #8384.

Change-Id: Id47e2c511678018c48686ec17ab03a59486bcb84
PiperOrigin-RevId: 250891894