Remove StartupOptions from globals
Demote from global-always-mutable state to sometimes-mutable state that's
passed around, making it clear what depends on it.

We should be able to mark StartupOptions const in many of the places we use it,
but the GetArgumentArray calls peppered throughout the code rely on non-const
GetServerJavabase. I'll figure out how to deal with that in another pass, either
by making GetServerJavabase const, lifting its use to a place that doesn't
matter, only computing args once, or some combination of such.

I stuck to passing a StartupOptions* in the code to keep the mass
"globals->options->" to "startup_options->" migration less interesting. By
reference would probably be better, but I think maybe for another go...

As usual, a downside to this is growing argument lists, but I think that's
better than guessing or having no enforcement of when a method might update
globals->options.

PiperOrigin-RevId: 251700379