Also record execution platforms in resolved files
The concept of execution platforms was introduced after the introduction
of the workspace resolved file. However, when it was added, no addition
was made to the resolved file to contain that information as well. Do so
now, to ensure the resolved file still can be used as a full replacement
for the workspace file.

Fixes #8523

Change-Id: I3e31845084393d13bc917071bc458ea8a78a4336
PiperOrigin-RevId: 252016507