Lift the part of Main after we're done mutating config into a new method
This way we can be (reasonably) confident that we're done mutating config, with
the exception of globals.

Also demoted install_base from globals, since it was a chunk of mutable state
used across the mutable/immutable split that was easy enough to remove.

Given we know that StartupOptions is const I'd like to update BlazeServer's ctor
to only retain the info it needs instead of all of StartupOptions, but I'll do
that in another go.

RELNOTES: None
PiperOrigin-RevId: 252107694