Fix broken debian packaging tests.
One test was clearly wrong. It had some extra characters in the expected text.

Another 2 checks were failing becuase of permissions on files within the
debian package. I could argue that we should fix the code and write the
files in 644 mode rather then 755, but this package is now deprecated,
so I do not want to waste effort fixing it here. It will be addressed in
https://github.com/bazelbuild/rules_pkg/issues/40

Closes #8601.

PiperOrigin-RevId: 252624041