Include intermediate output key as a typed parameter for ParallelVisitor.
This will allow for stricter type checking as well as the ability to store output keys other than just skykeys.

This CL is mostly a no-op but has one subtle change:  In AbstractEdgeVisitor (now AbstractTargetOuputtingVisitor), we acquire the semaphore only while fetching the packages but don't immediately process the results using the callback. We do this already in other visitors so this is more consistent, nor is it a good idea to have a semaphore lock around an unknown callback that may possibly try acquire keys too.

PiperOrigin-RevId: 253080306