Deflake test_build_fail_terse_summary.
The flakiness results from an actual race condition that makes it indeterminate whether CompletionFunction is called for the tests that depend on an artifact that fails to build.

As far as I understand, it should not be, since it depends on a SkyValue in error in a --nokeep_going build; however, according to the comment in AbstractParallelEvaluator.Evaluate.run() ("there is a race-like opportunity for the error to be used"), it's a known quirk of Skyframe.

printf-style debugging supports this theory: on flakes, CompletionFunction and AggregatingTestListener.targetComplete()) are called, on successful test runs, they are not.

Is sadness.

RELNOTES: None.
PiperOrigin-RevId: 253178762