Correct pluto repository directory in skylark_git_repository_test.sh,
the test for shallow option with a commit before the shallow date does not fail any more, as the callback without shallow options gets invoked under cover.
Add comment about that, keep the test for now to demonstrate the new behavior, and remove the test as superficial in the following commit.

Closes #8667.

PiperOrigin-RevId: 253940289