Add missing calls to testenv_tear_down.
The recently introduced testenv_tear_down method is a great improvement to overall test reliability, but we have to call it in all cases where we also call the original tear_down method.

Without these calls, the testenv.sh wouldn't clean up the workspace when a test failed, which could cause the next test method or the next set_up call of the test suite to fail, too, resulting in a chain reaction of failed tests. (I observed this with the bazel_sandboxing_test when trying to run tests inside a Docker container.)

RELNOTES: None.
PiperOrigin-RevId: 254734215