Add a Checksum class for plumbing HTTP downloader content checksums.
This is extracted from https://github.com/bazelbuild/bazel/pull/7208 ("Support subresource integrity format for `download()` checksums.")

I'm trying to revive that PR and rebase to master, but the replacement of `String sha256` through the HTTP downloader code is causing a lot of Git conflicts. I expect that portion of the change to require less discussion than the proposed content integrity changes, so I'm moving it to a separate PR to simplify review.

@dslomov or @aehlig, could one of you take a look? Once this is merged, I will rebase PR 7208 onto it.

Closes #8714.

PiperOrigin-RevId: 256188026