Windows: fix "insufficient resources" on Windows 7
On Windows 7, the subprocess cannot inherit
console handles via the attribute list
(CreateProcessW fails with
ERROR_NO_SYSTEM_RESOURCES).

However, if we:

- don't specify STARTF_USESTDHANDLES in
  STARTUPINFOW::dwFlags, and
- set CreateProcessW:bInheritHandles to FALSE,

then we disallow inheriting the current process'
open and inheritable handles.

However, if on top of that we also:

- don't specify CREATE_NEW_CONSOLE nor
  DETACHED_PROCESS in
  CreateProcessW:dwCreationFlags,

then, in this special case, the subprocess *does*
inherit the console handles, but nothing else
(i.e. none of the inheritable, open file handles
for example).

This is rather confusing, but seems to work and
do exactly what we need: for the subprocess to
inherit the console handles and nothing else.

https://github.com/bazelbuild/bazel/issues/8676

Closes #8793.

PiperOrigin-RevId: 256546818