Create canonical flag types for use within Bazel rules.
In an ideal world,  we would use the definitions from skylib. In the real world, we do not want to take a hard dependency on skylib in Bazel, so we will provide a set strictly for built-in Starlark based rules.

Note: These can not be generally used yet. They require --experimental_build_settings_api. That is planned to be turned on in 0.28.

Closes #8639.

PiperOrigin-RevId: 257327261